requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
cassandra-driver
kneed
